refactored and fixed functional tests junit version issue
