+++ src/functionalTest/java/uk/gov/hmcts/reform/em/hrs/functional/BaseTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.em.hrs;
+package uk.gov.hmcts.reform.em.hrs.functional;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -20,12 +20,12 @@ import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.em.EmTestConfig;
-import uk.gov.hmcts.reform.em.hrs.config.HrsAzureClient;
+import uk.gov.hmcts.reform.em.hrs.functional.config.HrsAzureClient;
 import uk.gov.hmcts.reform.em.hrs.model.CaseRecordingFile;
-import uk.gov.hmcts.reform.em.hrs.testutil.AuthTokenGeneratorConfiguration;
-import uk.gov.hmcts.reform.em.hrs.testutil.CcdAuthTokenGeneratorConfiguration;
-import uk.gov.hmcts.reform.em.hrs.testutil.ExtendedCcdHelper;
-import uk.gov.hmcts.reform.em.hrs.testutil.TestUtil;
+import uk.gov.hmcts.reform.em.hrs.functional.config.AuthTokenGeneratorConfiguration;
+import uk.gov.hmcts.reform.em.hrs.functional.config.CcdAuthTokenGeneratorConfiguration;
+import uk.gov.hmcts.reform.em.hrs.functional.util.ExtendedCcdHelper;
+import uk.gov.hmcts.reform.em.hrs.functional.util.TestUtil;
 import uk.gov.hmcts.reform.em.test.idam.IdamHelper;
 import uk.gov.hmcts.reform.em.test.retry.RetryRule;
 import uk.gov.hmcts.reform.em.test.s2s.S2sHelper;
@@ -39,7 +39,7 @@ import java.util.Optional;
 import javax.annotation.PostConstruct;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static uk.gov.hmcts.reform.em.hrs.testutil.ExtendedCcdHelper.HRS_TESTER;
+import static uk.gov.hmcts.reform.em.hrs.functional.util.ExtendedCcdHelper.HRS_TESTER;
 
 
 @SpringBootTest(classes = {
@@ -91,6 +91,9 @@ public abstract class BaseTest {
     @Autowired
     protected CoreCaseDataApi coreCaseDataApi;
 
+    @Autowired
+    private TestUtil testUtil;
+
     @PostConstruct
     public void init() {
         SerenityRest.useRelaxedHTTPSValidation();
@@ -137,7 +140,7 @@ public abstract class BaseTest {
             .get(recordingUrl);
     }
 
-    protected Response downloadRecording(String email, List<String> roles,Map<String, Object> caseData) {
+    protected Response downloadRecording(String email, List<String> roles, Map<String, Object> caseData) {
         @SuppressWarnings("unchecked")
         List<Map> segmentNodes = (ArrayList) caseData.getOrDefault("recordingFiles", new ArrayList());
 
@@ -214,7 +217,7 @@ public abstract class BaseTest {
         final Timestamp timestamp = new Timestamp(System.currentTimeMillis());
         return createRecordingSegmentPayload(
             FOLDER,
-            cvpContainerUrl + fileName,
+            testUtil.getCvpBlobContainerClient().getBlobContainerUrl() + "/" + fileName,
             this.fileName,
             FILE_EXT,
             SEGMENT,
